requirements.txt
aiosqlite==0.17.0
alembic==1.13.1
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==42.0.2
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.89.1
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
httpx-oauth==0.13.2
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.2
orjson==3.8.5
passlib==1.7.4
psycopg2==2.9.9
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.4
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==0.21.1
python-multipart==0.0.7
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.2
starlette==0.22.0
tortoise-orm==0.19.3
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
